+++ src/Controllers/E404Controller.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace App\Controllers;
 
-use App\Models\E404Model;
+use App\Models\NullModel;
 use Linna\Mvc\Controller;
 
 /**
@@ -22,9 +22,9 @@ class E404Controller extends Controller
     /**
      * Constructor.
      *
-     * @param E404Model $model
+     * @param NullModel $model
      */
-    public function __construct(E404Model $model)
+    public function __construct(NullModel $model)
     {
         //call parent construct
         parent::__construct($model);